Extract Options to own file
